feat(crypto): 가상화폐 모의투자 기능 entity 및 repository 개발-#55
- 거래를 하나의 event로, DB table의 append-only 방식
- offset 방식은 건너뛸 데이터까지 DB가 순차적으로 모두 읽고 버려야 하는 구조라, WHERE 구문으로 처리 가능한 ProcessorOffset Entity 추가

+++ tradelink/src/main/java/site/tradelink/tradelink/cryptocurrency/dto/TradeLogDto.java
@@ -8,7 +8,7 @@ import java.time.LocalDateTime;
 public record TradeLogDto (
         String        ticker,
         long          price,        // 체결가
-        long          quantity,     // 체결량
+        double          quantity,     // 체결량
         String        side,         // "BUY" | "SELL"
         LocalDateTime tradedAt
 ) {}
